requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
slowapi

langchain
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-groq

faiss-cpu
pypdf
python-dotenv